fix: 썸네일 이미지 삭제 시 빈 문자열 처리 추가

+++ src/main/java/com/example/gak/domain/session/service/SessionCommandService.java
@@ -517,7 +517,7 @@ public class SessionCommandService {
 			String keyName = amazonS3Manager.generateSessionThumbnailKeyName();
 			String imageUrl = amazonS3Manager.uploadFile(keyName, image);
 
-			if (sessionRoom.getThumbnailImageUrl() != null) {
+			if (sessionRoom.getThumbnailImageUrl() != null && !sessionRoom.getThumbnailImageUrl().isEmpty()) {
 				amazonS3Manager.deleteFile(sessionRoom.getThumbnailImageUrl());
 			}
